printPowersOfN

+++ Power.java
@@ -8,7 +8,7 @@ public class Power {
     	public static void printPowersOfN(int num1, int count) {
     		for(int i = 1; i <=count; i++) {
     			double result= Math.pow(num1, i);
-    			System.out.println(result);
+    			System.out.println((int)(result));
     		}
     	}
